weapon

+++ Assets/Script/player/Player/heal/IDamageable.cs
@@ -2,6 +2,6 @@ namespace Script.player.Player.heal
 {
     public interface IDamageable
     {
-        void Apply(float damage);
+        void Apply(int damage);
     }
 }
